testmessage: move TestMessage into it's own module
